+++ tensorflow/cppbuild.sh
@@ -12,6 +12,7 @@ export USE_DEFAULT_PYTHON_LIB_PATH=1
 export TF_NEED_CUDA=0
 export TF_NEED_GCP=0
 export TF_NEED_HDFS=0
+export TF_NEED_OPENCL=0
 export TF_CUDA_VERSION=8.0
 export TF_CUDNN_VERSION=5
 export GCC_HOST_COMPILER_PATH=$(which gcc)
@@ -19,9 +20,9 @@ export CUDA_TOOLKIT_PATH=/usr/local/cuda
 export CUDNN_INSTALL_PATH=$CUDA_TOOLKIT_PATH
 export TF_CUDA_COMPUTE_CAPABILITIES=3.0
 
-TENSORFLOW_VERSION=0.11.0
+TENSORFLOW_VERSION=0.12.1
 
-download https://github.com/tensorflow/tensorflow/archive/v$TENSORFLOW_VERSION.tar.gz tensorflow-$TENSORFLOW_VERSION.tar.gz
+download https://github.com/tensorflow/tensorflow/archive/$TENSORFLOW_VERSION.tar.gz tensorflow-$TENSORFLOW_VERSION.tar.gz
 
 mkdir -p $PLATFORM
 cd $PLATFORM
@@ -32,7 +33,8 @@ tar --totals -xzf ../tensorflow-$TENSORFLOW_VERSION.tar.gz
 # Assumes Bazel is available in the path: http://bazel.io/docs/install.html
 cd tensorflow-$TENSORFLOW_VERSION
 
-patch -Np1 < ../../../tensorflow-$TENSORFLOW_VERSION.patch
+sed -i s/zlib-1.2.8/zlib-1.2.10/ tensorflow/workspace.bzl
+sed -i s/36658cb768a54c1d4dec43c3116c27ed893e88b02ecfcb44f2166f9c0b7f2a0d/8d7e9f698ce48787b6e1c67e6bff79e487303e66077e25cb9784ac8835978017/ tensorflow/workspace.bzl
 
 case $PLATFORM in
 	android-arm)