+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 
- * Upgrade presets for LLVM 3.9.1 ([pull #343](https://github.com/bytedeco/javacpp-presets/pull/343))
+ * Upgrade presets for libdc1394 2.2.5, LLVM 3.9.1 ([pull #343](https://github.com/bytedeco/javacpp-presets/pull/343)), TensorFlow 0.12.1
  * Add presets for libfreenect2 ([pull #340](https://github.com/bytedeco/javacpp-presets/pull/340))
- * Upgrade presets for libdc1394 2.2.5
  * Fix the `FlyCapture2` module for some versions on Windows ([issue #337](https://github.com/bytedeco/javacpp-presets/issues/337))
  * Add functions missing from the presets of MXNet ([issue #332](https://github.com/bytedeco/javacpp-presets/issues/332))
  * Add presets for the `text` module of OpenCV 3.1 ([pull #333](https://github.com/bytedeco/javacpp-presets/pull/333))